refactor: file restructuring
