+++ src/components/TeacherComponents/LessonOverviewSideNavAnchorLinks/LessonOverviewSideNavAnchorLinks.tsx
@@ -12,17 +12,73 @@ import {
   LessonOverviewDownloadAllButton,
   LessonOverviewDownloadAllButtonProps,
 } from "../LessonOverviewDownloadAllButton/LessonOverviewDownloadAllButton";
+import { LessonResource } from "../../../app/(core)/programmes/[subjectPhaseSlug]/units/[unitSlug]/lessons/[lessonSlug]/Components/getLessonResources";
+import { DISPLAY_TITLES_BY_RESOURCE } from "../../../app/(core)/programmes/[subjectPhaseSlug]/units/[unitSlug]/lessons/[lessonSlug]/Components/lessonResourceConstants";
 
 type LessonOverviewSideNavAnchorLinksProps = {
   contentRestricted: boolean;
   currentSectionId: string | null;
-  links: {
-    label: string;
-    anchorId: string;
-    subheading?: string;
-  }[];
+  links: SideNavLink[];
   downloadAllButtonProps: LessonOverviewDownloadAllButtonProps;
 };
+
+export type SideNavLink = {
+  label: string;
+  anchorId: string;
+  subheading?: string;
+};
+
+export function getSideNavLinksFromResources(
+  resources: LessonResource[],
+): SideNavLink[] {
+  const hasStarterQuiz = resources.some(
+    (r) => r.resourceType === "starter-quiz",
+  );
+  const hasExitQuiz = resources.some((r) => r.resourceType === "exit-quiz");
+  const hasBothQuizzes = hasStarterQuiz && hasExitQuiz;
+
+  const links: SideNavLink[] = [];
+  let quizLinkAdded = false;
+
+  for (const resource of resources) {
+    const isQuizResource =
+      resource.resourceType === "starter-quiz" ||
+      resource.resourceType === "exit-quiz";
+
+    if (isQuizResource) {
+      // Only add the quiz link once, at the position of the first quiz
+      if (!quizLinkAdded) {
+        if (hasBothQuizzes) {
+          links.push({
+            label: "Quizzes",
+            anchorId: resource.anchorId, // "quiz" for first quiz when both exist
+            subheading: `${DISPLAY_TITLES_BY_RESOURCE["starter-quiz"]} \n${DISPLAY_TITLES_BY_RESOURCE["exit-quiz"]}`,
+          });
+        } else if (hasStarterQuiz) {
+          links.push({
+            label: "Quizzes",
+            anchorId: resource.anchorId, // "starter-quiz"
+            subheading: DISPLAY_TITLES_BY_RESOURCE["starter-quiz"],
+          });
+        } else if (hasExitQuiz) {
+          links.push({
+            label: "Quizzes",
+            anchorId: resource.anchorId, // "exit-quiz"
+            subheading: DISPLAY_TITLES_BY_RESOURCE["exit-quiz"],
+          });
+        }
+        quizLinkAdded = true;
+      }
+    } else {
+      links.push({
+        label: resource.title,
+        anchorId: resource.resourceType,
+      });
+    }
+  }
+
+  return links;
+}
 
 const LessonOverviewSideNavAnchorLinks: FC<
   LessonOverviewSideNavAnchorLinksProps
